Refactored all the CLI code in a single file
